+++ src/main/java/miyucomics/hexical/features/lamps/LampCastEnv.kt
@@ -3,16 +3,24 @@ package miyucomics.hexical.features.lamps
 import at.petrak.hexcasting.api.casting.ParticleSpray
 import at.petrak.hexcasting.api.casting.eval.CastResult
 import at.petrak.hexcasting.api.casting.eval.env.PackagedItemCastEnv
+import at.petrak.hexcasting.api.casting.iota.Iota
+import at.petrak.hexcasting.api.casting.iota.IotaType
 import at.petrak.hexcasting.api.casting.iota.NullIota
 import at.petrak.hexcasting.api.pigment.FrozenPigment
+import at.petrak.hexcasting.api.utils.putCompound
 import at.petrak.hexcasting.xplat.IXplatAbstractions
+import net.minecraft.item.ItemStack
 import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.util.Hand
 
-open class LampCastEnv(caster: ServerPlayerEntity, castingHand: Hand, private val finale: Boolean) : PackagedItemCastEnv(caster, castingHand) {
+class LampCastEnv(caster: ServerPlayerEntity, castingHand: Hand, val finale: Boolean, val stack: ItemStack) : PackagedItemCastEnv(caster, castingHand) {
 	override fun produceParticles(particles: ParticleSpray, pigment: FrozenPigment) {}
 	override fun postExecution(result: CastResult) = super.postExecution(result.copy(cast = NullIota()))
 	override fun getCastingHand(): Hand = this.castingHand
 	override fun getPigment(): FrozenPigment = IXplatAbstractions.INSTANCE.getPigment(this.caster)
+
 	fun getFinale() = finale
+	fun setInternalIota(iota: Iota) {
+		stack.orCreateNbt.putCompound("storage", IotaType.serialize(iota))
+	}
 }