+++ src/main/java/miyucomics/hexical/features/lamps/OpSetLampStorage.kt
@@ -5,17 +5,17 @@ import at.petrak.hexcasting.api.casting.castables.SpellAction
 import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.iota.Iota
 
-object OpSetHandLampStorage : SpellAction {
+object OpSetLampStorage : SpellAction {
 	override val argc = 1
 	override fun execute(args: List<Iota>, env: CastingEnvironment): SpellAction.Result {
-		if (env !is HandLampCastEnv)
-			throw NeedsHandLampMishap()
+		if (env !is LampCastEnv)
+			throw NeedsLampMishap()
 		return SpellAction.Result(Spell(args[0]), 0, listOf())
 	}
 
 	private data class Spell(val iota: Iota) : RenderedSpell {
 		override fun cast(env: CastingEnvironment) {
-			(env as HandLampCastEnv).setInternalIota(iota)
+			(env as LampCastEnv).setInternalIota(iota)
 		}
 	}
 }