+++ src/main/java/miyucomics/hexical/features/player/PlayerManager.kt
@@ -4,7 +4,6 @@ import miyucomics.hexical.features.evocation.EvocationField
 import miyucomics.hexical.features.evocation.EvocationTicker
 import miyucomics.hexical.features.item_cache.PlayerItemCacheField
 import miyucomics.hexical.features.item_cache.PlayerItemCacheTicker
-import miyucomics.hexical.features.lamps.ArchLampField
 import miyucomics.hexical.features.lesser_sentinels.LesserSentinelField
 import miyucomics.hexical.features.media_log.MediaLogField
 import miyucomics.hexical.features.player.types.PlayerField
@@ -26,7 +25,6 @@ class PlayerManager {
 	private fun registerTicker(ticker: PlayerTicker) { tickers.add(ticker) }
 
 	init {
-		registerField(ArchLampField())
 		registerField(EvocationField())
 		registerField(KeybindField())
 		registerField(MediaLogField())